# docs/swagger.yaml
openapi: 3.0.4
info:
  title: E-commerce API
  description: |
    Complete e-commerce API with user authentication, product management, 
    shopping cart, orders, payments, and admin functionality.
    
    ## Authentication
    - JWT Bearer token authentication
    - Email verification required for full access
    - Google OAuth support
    
    ## Roles
    - **user**: Regular customer permissions
    - **admin**: Full administrative access
    
    ## Rate Limiting
    - 1000 requests per hour per IP
    - 100 requests per minute for authenticated users
  version: 1.0.0
  termsOfService: https://example.com/terms
  contact:
    name: Abdulrahman Hegazy
    email: a.0hegazy@proton.me
...
servers:
  - url: http://localhost:5000/api/v1
    description: Local development server
  - url: https://ecommerce-express-boilerplate.onrender.com/api/v1
    description: Production server

# Global security - applies to all endpoints
security:
  - bearerAuth: []
...
paths:
...
  /google-login:
    post:
      tags:
        - Authentication
      summary: Google OAuth login
      description: Authenticate user using Google OAuth credentials
      operationId: googleLogin
      requestBody:
        required: true
        description: Google OAuth user data
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleLoginRequest'
      responses:
        '200':
          description: Google login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Successful Google login
                  value:
                    status: success
                    data:
                      user:
                        id: "64c9b8d3a7e9a2d8f0c1b234"
                        name: "John Doe"
                        email: "john@gmail.com"
                        role: "user"
                        avatar: "https://lh3.googleusercontent.com/..."
                        emailVerified: true
                        isActive: true
                        createdAt: "2025-10-15T12:00:00.000Z"
                        updatedAt: "2025-10-15T12:30:00.000Z"
                      accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      message: "Google login successful"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Generate new access token using refresh token (from cookie or body)
      operationId: refreshToken
      requestBody:
        required: false
        description: Optional refresh token (if not provided in cookie)
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Access token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Token refreshed
                  value:
                    status: success
                    data:
                      accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      message: "Access token refreshed successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /verify-email:
    post:
      tags:
        - Authentication
      summary: Verify email address
      description: Verify user's email address using verification token
      operationId: verifyEmail
      security: []  # No auth required for email verification
      requestBody:
        required: true
        description: Email verification token
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyEmailRequest'
      responses:
        '200':
          description: Email verified successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Email verified
                  value:
                    status: success
                    data:
                      message: "Email verified successfully! Welcome to our platform."
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /resend-verification:
    post:
      tags:
        - Authentication
      summary: Resend verification email
      description: Resend email verification link to user's email address
      operationId: resendVerificationEmail
      security: []  # No auth required
      requestBody:
        required: true
        description: User email address
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequest'
      responses:
        '200':
          description: Verification email sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Verification email sent
                  value:
                    status: success
                    data:
                      message: "Verification email sent successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /forgot-password:
    post:
      tags:
        - Authentication
      summary: Request password reset
      description: Send password reset link to user's email address
      operationId: forgotPassword
      security: []  # No auth required
      requestBody:
        required: true
        description: User email address
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset link sent if account exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Reset email sent
                  value:
                    status: success
                    data:
                      message: "If an account with that email exists, a password reset link has been sent."
        '500':
          $ref: '#/components/responses/ServerError'
  /reset-password:
    post:
      tags:
        - Authentication
      summary: Reset password
      description: Reset user password using valid reset token
      operationId: resetPassword
      security: []  # No auth required
      requestBody:
        required: true
        description: Reset token and new password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Password reset successful
                  value:
                    status: success
                    data:
                      user:
                        id: "64c9b8d3a7e9a2d8f0c1b234"
                        name: "John Doe"
                        email: "john@example.com"
                        role: "user"
                        emailVerified: true
                        isActive: true
                      accessToken: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                      message: "Password reset successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /change-password:
    patch:
      tags:
        - Authentication
      summary: Change password
      description: Change password for authenticated user
      operationId: changePassword
      requestBody:
        required: true
        description: Current and new password
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Password changed
                  value:
                    status: success
                    data:
                      message: "Password changed successfully"
        '400':
          description: Bad request - Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                incorrectPassword:
                  summary: Incorrect current password
                  value:
                    status: fail
                    data:
                      message: "Current password is incorrect"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /profile:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Retrieve authenticated user's profile information
      operationId: getProfile
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Profile retrieved
                  value:
                    status: success
                    data:
                      user:
                        id: "64c9b8d3a7e9a2d8f0c1b234"
                        name: "John Doe"
                        email: "john@example.com"
                        role: "user"
                        avatar: "https://via.placeholder.com/150/cccccc/969696?text=User"
                        phoneNumber: "+1234567890"
                        emailVerified: true
                        isActive: true
                        address:
                          street: "123 Main St"
                          city: "New York"
                          state: "NY"
                          zipCode: "10001"
                          country: "US"
                        preferences:
                          newsletter: true
                          notifications: false
                        createdAt: "2025-10-15T12:00:00.000Z"
                        updatedAt: "2025-10-15T12:30:00.000Z"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    patch:
      tags:
        - Authentication
      summary: Update user profile
      description: Update authenticated user's profile information
      operationId: updateProfile
      requestBody:
        required: true
        description: Profile data to update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                success:
                  summary: Profile updated
                  value:
                    status: success
                    data:
                      user:
                        id: "64c9b8d3a7e9a2d8f0c1b234"
                        name: "John Smith"
                        email: "john@example.com"
                        role: "user"
                        phoneNumber: "+1234567890"
                        address:
                          street: "456 Oak Ave"
                          city: "Los Angeles"
                          state: "CA"
                          zipCode: "90210"
                          country: "US"
                        preferences:
                          newsletter: false
                          notifications: true
                      message: "Profile updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication. For admin endpoints, user must have 'admin' role.
...
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: fail
            data:
              message: "Unauthorized. Please log in."
...
    BadRequest:
      description: Bad request - Invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: fail
            data:
              message: "Invalid input data"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            status: fail
            data:
              message: "Resource not found"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            status: error
            data: "Internal Server Error"
  schemas:
    ApiResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, fail, error]
          example: success
        data:
          oneOf:
            - type: object
              description: Contains response data in success and fail cases
              properties:
                message:
                  type: string
                  example: "Successfully completed action"
                accessToken:
                  type: string
                  example: "eyJhbGciOiJIUzI1NiIsInR5..."
                user:
                  $ref: '#/components/schemas/User'
            - type: string
              description: Contains error message or additional info in error cases
              example: "Internal server error"
            - type: object
              nullable: true
              description: If no data is available, null can be provided
              example: null
      required:
        - status
        - data
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          enum: [error]
          example: error
        data:
          type: string
          example: "Internal Server Error"
...
    GoogleLoginRequest:
      type: object
      required:
        - googleId
        - email
        - name
      properties:
        googleId:
          type: string
          example: "1234567890"
        email:
          type: string
          format: email
          example: "demo@gmail.com"
        name:
          type: string
          example: "Demo user"
        picture:
          type: string
          format: uri
          example: "https://lh3.googleusercontent.com/..."
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 50
          example: "John Smith"
        phoneNumber:
          type: string
          example: "+1234567890"
        address:
          type: object
          properties:
            street:
              type: string
              example: "456 Oak Ave"
            city:
              type: string
              example: "Los Angeles"
            state:
              type: string
              example: "CA"
            zipCode:
              type: string
              example: "90210"
            country:
              type: string
              default: "US"
              example: "US"
        preferences:
          type: object
          properties:
            newsletter:
              type: boolean
              default: true
              example: false
            notifications:
              type: boolean
              default: true
              example: true
    VerifyEmailRequest:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          example: "email_verification_token_123"
    ResendVerificationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: "john@example.com"
    RefreshTokenRequest:
      type: object
      properties:
        refreshToken:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: "john@example.com"
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
      properties:
        token:
          type: string
          example: "password_reset_token_123"
        password:
          type: string
          format: password
          minLength: 6
          example: "newSecurePass123@"
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
          format: password
          example: "oldPassword123@"
        newPassword:
          type: string
          format: password
          minLength: 6
          example: "newSecurePassw456@"
...
    User:
      type: object
      properties:
        id:
          type: string
          example: "64c9b8d3a7e9a2d8f0c1b234"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [user, admin]
          example: "user"
        avatar:
          type: string
          format: uri
          example: "https://via.placeholder.com/150/cccccc/969696?text=User"
        phoneNumber:
          type: string
          example: "+1234567890"
        emailVerified:
          type: boolean
          example: true
        isActive:
          type: boolean
          example: true
        address:
          $ref: '#/components/schemas/Address'
        preferences:
          $ref: '#/components/schemas/UserPreferences'
        createdAt:
          type: string
          format: date-time
          example: "2025-10-15T12:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2025-10-15T12:30:00.000Z"
    Address:
      type: object
      properties:
        street:
          type: string
          example: "123 Main St"
        city:
          type: string
          example: "New York"
        state:
          type: string
          example: "NY"
        zipCode:
          type: string
          example: "10001"
        country:
          type: string
          default: "US"
          example: "US"
    UserPreferences:
      type: object
      properties:
        newsletter:
          type: boolean
          default: true
          example: true
        notifications:
          type: boolean
          default: true
          example: false
...